refactor: Json decoder extracted to a separate class
