+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/projectdetails/model.scala
@@ -24,18 +24,10 @@ import eu.timepit.refined.api.Refined
 import eu.timepit.refined.auto._
 import eu.timepit.refined.collection.NonEmpty
 import eu.timepit.refined.numeric.Positive
-import io.circe.Encoder
-import io.circe.literal._
-import io.renku.config.renku
 import io.renku.graph.model.projects.{DateCreated, Description, Id, Keyword, Name, Path, Visibility}
 import io.renku.graph.model.{SchemaVersion, persons}
-import io.renku.http.rest.Links.{Link, Rel, _links}
-import io.renku.json.JsonOps._
-import io.renku.knowledgegraph.datasets.ProjectDatasetsEndpoint
 import io.renku.tinytypes._
 import io.renku.tinytypes.constraints._
-import io.renku.tinytypes.json.TinyTypeEncoders._
-import model.Permissions._
 import model.Statistics._
 
 import java.net.{MalformedURLException, URL}
@@ -198,93 +190,5 @@ private object model {
 
     final class ReadmeUrl private (val value: String) extends AnyVal with StringTinyType
     implicit object ReadmeUrl extends TinyTypeFactory[ReadmeUrl](new ReadmeUrl(_)) with Url[ReadmeUrl]
-  }
-
-  private[projectdetails] implicit def encoder(implicit renkuApiUrl: renku.ApiUrl): Encoder[Project] =
-    Encoder.instance[Project] { project =>
-      json"""{
-        "identifier": ${project.id},
-        "path":       ${project.path},
-        "name":       ${project.name},
-        "visibility": ${project.visibility},
-        "created":    ${project.created},
-        "updatedAt":  ${project.updatedAt},
-        "urls":       ${project.urls},
-        "forking":    ${project.forking},
-        "keywords":   ${project.keywords.toList.sorted},
-        "starsCount": ${project.starsCount},
-        "permissions":${project.permissions},
-        "statistics": ${project.statistics}
-      }""" deepMerge _links(
-        Link(Rel.Self        -> Endpoint.href(renkuApiUrl, project.path)),
-        Link(Rel("datasets") -> ProjectDatasetsEndpoint.href(renkuApiUrl, project.path))
-      ).addIfDefined("description" -> project.maybeDescription)
-        .addIfDefined("version" -> project.maybeVersion)
-    }
-
-  private implicit lazy val creatorEncoder: Encoder[Creator] = Encoder.instance[Creator] { creator =>
-    json"""{
-      "name":  ${creator.name}
-    }""" addIfDefined ("email" -> creator.maybeEmail)
-  }
-
-  private implicit lazy val urlsEncoder: Encoder[Urls] = Encoder.instance[Urls] { urls =>
-    json"""{
-      "ssh":    ${urls.ssh},
-      "http":   ${urls.http},
-      "web":    ${urls.web}
-    }""" addIfDefined ("readme" -> urls.maybeReadme)
-  }
-
-  private implicit lazy val forkingEncoder: Encoder[Forking] = Encoder.instance[Forking] { forks =>
-    json"""{
-      "forksCount": ${forks.forksCount}
-    }""" addIfDefined ("parent" -> forks.maybeParent)
-  }
-
-  private implicit lazy val parentProjectEncoder: Encoder[ParentProject] = Encoder.instance[ParentProject] { parent =>
-    json"""{
-      "path":    ${parent.path},
-      "name":    ${parent.name},
-      "created": ${parent.created}
-    }"""
-  }
-
-  private implicit lazy val creationEncoder: Encoder[Creation] = Encoder.instance[Creation] { created =>
-    json"""{
-      "dateCreated": ${created.date}
-    }""" addIfDefined ("creator" -> created.maybeCreator)
-  }
-
-  private implicit lazy val permissionsEncoder: Encoder[Permissions] = Encoder.instance[Permissions] {
-    case ProjectAndGroupPermissions(projectAccessLevel, groupAccessLevel) => json"""{
-      "projectAccess": ${projectAccessLevel.accessLevel},
-      "groupAccess":   ${groupAccessLevel.accessLevel}
-    }"""
-    case ProjectPermissions(accessLevel) => json"""{
-      "projectAccess": ${accessLevel.accessLevel}
-    }"""
-    case GroupPermissions(accessLevel) => json"""{
-      "groupAccess": ${accessLevel.accessLevel}
-    }"""
-  }
-
-  private implicit lazy val accessLevelEncoder: Encoder[AccessLevel] = Encoder.instance[AccessLevel] { level =>
-    json"""{
-      "level": {
-        "name":  ${level.name.value},
-        "value": ${level.value.value}
-      }
-    }"""
-  }
-
-  private implicit lazy val statisticsEncoder: Encoder[Statistics] = Encoder.instance[Statistics] { stats =>
-    json"""{
-      "commitsCount":     ${stats.commitsCount},
-      "storageSize":      ${stats.storageSize},
-      "repositorySize":   ${stats.repositorySize},
-      "lfsObjectsSize":   ${stats.lsfObjectsSize},
-      "jobArtifactsSize": ${stats.jobArtifactsSize}
-    }"""
   }
 }
